♻️ refactor(test): reorganize test infrastructure and improve patterns
- Rename test doubles to follow .fake.ts and .stub.ts conventions
- Move test builders to test-support/builders/ directory
- Move fake implementations to test-utils/fakes/ with barrel exports
- Create testkit/index.ts for centralized test utility exports
- Add AuthProviderFactory.prod.ts for production-only auth provider
- Update import paths across test files to use new structure
- Improve TypeScript configuration for test utilities
