+++ apps/api/src/infra/auth/AuthProvider.stub.ts
