+++ apps/api/src/test-support/builders/QuizSessionBuilder.ts
@@ -4,9 +4,9 @@
  */
 
 import { TestClock, testIds } from '@api/test-support';
-import type { QuestionId, UserId } from '../value-objects/Ids';
-import { QuizConfig } from '../value-objects/QuizConfig';
-import { QuizSession } from './QuizSession';
+import { QuizSession } from '../../features/quiz/domain/aggregates/QuizSession';
+import type { QuestionId, UserId } from '../../features/quiz/domain/value-objects/Ids';
+import { QuizConfig } from '../../features/quiz/domain/value-objects/QuizConfig';
 
 interface QuizSessionBuilderProps {
   userId?: UserId;
