+++ apps/api/test-utils/fakes/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Test Fakes Barrel Export
+ * @fileoverview Centralized exports for all fake implementations
+ */
+
+export { FakeQuizRepository } from './FakeQuizRepository';
+export { FakeUnitOfWork, FakeUnitOfWorkFactory, withFakeUnitOfWork } from './FakeUnitOfWork';
+export { FakeUserRepository } from './FakeUserRepository';
