+++ apps/api/tests/fakes/index.ts
@@ -1,15 +1,8 @@
 /**
- * Test Fakes Barrel Export
- * @fileoverview Centralized exports for all fake implementations
+ * Test fakes barrel export
+ * @fileoverview Re-exports all fake implementations
  */
 
-export { FakeQuizRepository } from './FakeQuizRepository';
-export {
-  FakeUnitOfWork,
-  FakeUnitOfWorkFactory,
-  withFakeUnitOfWork,
-} from './FakeUnitOfWork';
-export { FakeUserRepository } from './FakeUserRepository';
-
-// TODO: Add more fake implementations as needed
-// export { FakeEmailService } from './FakeEmailService';
+export { FakeQuizRepository } from '../../test-utils/fakes/FakeQuizRepository';
+export { FakeUnitOfWork } from '../../test-utils/fakes/FakeUnitOfWork';
+export { FakeUserRepository } from '../../test-utils/fakes/FakeUserRepository';
